growth_mate_project-master/requirements.txt
asgiref==3.8.1
Django==5.1.7
dotenv==0.9.9
pillow==11.1.0
python-dotenv==1.0.1
six==1.17.0
sqlparse==0.5.3
tzdata==2025.1
google-generativeai==0.3.2
